refactor(web-search): split executor into provider, normalization, and formatting modules
